paint input system

+++ Engine/Source/Vector2.h
@@ -8,7 +8,7 @@ struct Vector2
 
 	Vector2() = default;
 	Vector2(float x, float y) { this->x = x; this->y = y; };
-	Vector2(int x, int y) { this->x = x; this->y = y; };
+	Vector2(int x, int y) { this->x = (float)x; this->y = (float)y; };
 
 	//Vector2 add(Vector2 v) { return Vector2{x + v.x, y + v.y}; }
